WebSockets/tests/submissions/Opera/opening-handshake/001.html

<!doctype html>
<title>WebSockets: invalid handshake</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
setup({timeout:10010});

async_test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/ws/invalid');
  ws.onclose = t.step_func(function(e) {
    assert_false(e.wasClean);
    ws.onclose = t.step_func(function() {assert_unreached()});
    setTimeout(t.step_func(function() {t.done();}), 50)
  });
   ws.onmessage = ws.onopen = t.step_func(function() {assert_unreached()});
}, null, {timeout:10000});
</script>
